# shared/openApi.yaml
openapi: 3.0.3
info:
  title: Swagger - OpenAPI 3.0
  contact:
    email: nikita@ttnr.me
  version: 1.0.0
servers:
  - url: http://api.template.de/
    description: Dev server
  - url: https://euw.backend.example.com/
    description: Prod server
...
paths:
...
  /products:
    post:
      tags:
        - product
      summary: Create a new product
      requestBody:
        $ref: '#/components/requestBodies/Product'
      responses:
        '201':
          $ref: '#/components/responses/Product'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
    get:
      tags:
        - product
      summary: Retrieve all products
      responses:
        '200':
          $ref: '#/components/responses/ProductList'
        '401':
          $ref: '#/components/responses/Error'
  /products/{productId}:
    get:
      tags:
        - product
      summary: Retrieve a single product
      parameters:
        - name: productId
          in: path
          required: true
          description: The ID of the product to retrieve
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/ProductDetail'
        '404':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
  /products/{productId}/rent:
    post:
      tags:
        - product
      summary: Rent a product
      parameters:
        - name: productId
          in: path
          required: true
          description: The ID of the product to rent
          schema:
            type: string
      requestBody:
        $ref: '#/components/requestBodies/RentProductFormular'
      responses:
        '200':
          $ref: '#/components/responses/RentProductResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    Error:
      description: Error messages
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Product:
      description: Product details
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Product'
    ProductList:
      description: List of products
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Product'
    ProductDetail:
      description: Detailed product information
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/Product'
              - $ref: '#/components/schemas/ProductPublic'
    RentProductResponse:
      description: Product rented successfully
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RentProductConfirmation'
  requestBodies:
...
    Product:
      description: Product details
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Product'
    RentProductFormular:
      description: Rental formular
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RentProductFormular'
...
  schemas:
...
    Error:
      type: object
      properties:
        warningMessages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        errorMessages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
    Message:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
...
    Product:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the product
        name:
          type: string
          description: Name of the product
        description:
          type: string
          description: Description of the product
        images:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                description: Unique identifier for the image
              name:
                type: string
                description: Name of the image
              data:
                type: string
                format: byte
                description: Base64 encoded image data
            required:
              - id
              - name
              - data
          description: List of base64 encoded image data with IDs and names
        documents:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                description: Unique identifier for the document
              name:
                type: string
                description: Name of the document
              data:
                type: string
                format: byte
                description: Base64 encoded document data
            required:
              - id
              - name
              - data
          description: List of base64 encoded document data with IDs and names
        location:
          type: string
          description: Location ID of the product
        pricing:
          type: object
          properties:
            price:
              type: number
              format: float
              description: Price of the product
            deposit:
              type: number
              format: float
              description: Deposit amount for the product
          required:
            - price
          description: Pricing details including price and deposit
        dynamicAttributes:
          type: object
          additionalProperties: true
          description: Dynamic attributes for the product
      required:
        - name
        - description
        - pricing
        - location
    ProductPublic:
      allOf:
        - $ref: '#/components/schemas/Product'
        - type: object
          properties:
            id:
              type: string
              description: Unique identifier for the product
            isRented:
              type: boolean
              description: Indicates whether the product is currently rented
            isCurrentUserRenter:
              type: boolean
              description: Indicates whether the current user is the renter of the product
            qrCodeUrl:
              type: string
              description: URL of the generated QR code
            renterInfo:
              type: object
              properties:
                rentalStartDate:
                  type: integer
                  format: int64
                  description: Unix timestamp for the start date of the rental period
                rentalEndDate:
                  type: integer
                  format: int64
                  description: Unix timestamp for the end date of the rental period
                expectedDuration:
                  type: string
                  description: Expected duration of the rental
              description: Information about the rental period
...
    RentProductFormular:
      type: object
      properties:
        rentalStartDate:
          type: integer
          format: int64
          description: Unix timestamp for the start date of the rental period
        rentalEndDate:
          type: integer
          format: int64
          description: Unix timestamp for the end date of the rental period
        additionalNotes:
          type: string
          maxLength: 1000
          description: Free text for additional notes or special instructions
        locationId:
          type: string
          description: Identifier for the location where the product will be picked up
        paymentMethodId:
          type: string
          description: Identifier for the payment method used for the rental
        dynamicAttributes:
          type: object
          additionalProperties: true
          description: Dynamic attributes for the rental request
      required:
        - rentalStartDate
        - rentalEndDate
        - locationId
        - paymentMethodId
    RentProductConfirmation:
      type: object
      properties:
        rentContractId:
          type: string
          description: Unique identifier for the rent contract
...