# src/backend/open-api/models/model_message.go
/*
 * Swagger - OpenAPI 3.0
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 * Contact: nikita@ttnr.me
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type Message struct {
	Code string `json:"code,omitempty"`

	Message string `json:"message,omitempty"`
}

// AssertMessageRequired checks if the required fields are not zero-ed
func AssertMessageRequired(obj Message) error {
	return nil
}

// AssertMessageConstraints checks if the values respects the defined constraints
func AssertMessageConstraints(obj Message) error {
	return nil
}