# src/backend/open-api/models/model_product_public.go
/*
 * Swagger - OpenAPI 3.0
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 * Contact: nikita@ttnr.me
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type ProductPublic struct {

	// Name of the product
	Name string `json:"name"`

	// Description of the product
	Description string `json:"description"`

	// List of base64 encoded image data with IDs and names
	Images []ProductImagesInner `json:"images,omitempty"`

	// List of base64 encoded document data with IDs and names
	Documents []ProductDocumentsInner `json:"documents,omitempty"`

	// Location ID of the product
	Location string `json:"location"`

	Pricing ProductPricing `json:"pricing"`

	// Dynamic attributes for the product
	DynamicAttributes map[string]interface{} `json:"dynamicAttributes,omitempty"`

	// Unique identifier for the product
	ID string `json:"id,omitempty"`

	// Indicates whether the product is currently rented
	IsRented bool `json:"isRented,omitempty"`

	// Indicates whether the current user is the renter of the product
	IsCurrentUserRenter bool `json:"isCurrentUserRenter,omitempty"`

	// URL of the generated QR code
	QrCodeUrl string `json:"qrCodeUrl,omitempty"`

	RenterInfo ProductPublicAllOfRenterInfo `json:"renterInfo,omitempty"`
}

// AssertProductPublicRequired checks if the required fields are not zero-ed
func AssertProductPublicRequired(obj ProductPublic) error {
	elements := map[string]interface{}{
		"name":        obj.Name,
		"description": obj.Description,
		"location":    obj.Location,
		"pricing":     obj.Pricing,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	for _, el := range obj.Images {
		if err := AssertProductImagesInnerRequired(el); err != nil {
			return err
		}
	}
	for _, el := range obj.Documents {
		if err := AssertProductDocumentsInnerRequired(el); err != nil {
			return err
		}
	}
	if err := AssertProductPricingRequired(obj.Pricing); err != nil {
		return err
	}
	if err := AssertProductPublicAllOfRenterInfoRequired(obj.RenterInfo); err != nil {
		return err
	}
	return nil
}

// AssertProductPublicConstraints checks if the values respects the defined constraints
func AssertProductPublicConstraints(obj ProductPublic) error {
	return nil
}